% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.achieveGap}
\alias{print.achieveGap}
\title{Print Method for achieveGap Objects}
\usage{
\method{print}{achieveGap}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"achieveGap"}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Method for achieveGap Objects
}
